Add simplified geometry.
Processed with
    ogr2ogr -progress -f "ESRI Shapefile" pd_a.simp.shp pd_a.shp -nlt POLYGON -simplify 1.0

Some checking in QGIS indicates that there don't seem to be any
major changes in the geometry, even for fine boundaries like in
city downtowns. Since the OGR simplification algorithm does not preserve
layer topology, it's best to be cautious -- but this still gives a 4x
reduction in geometry size.